Use the QT_NO_KEYWORDS define
Since 2.25.x glib defines 'signals' as a member of a struct in a
public header. This breaks when compiling Qt because it #defines
signals to 'protected'. To work around this ClutterQt now defines the
QT_NO_KEYWORDS macro which stops the Qt headers from doing that. I
guess this also means that all applications using ClutterQt would have
to define that too.

There is a Gnome bug about it here:

https://bugzilla.gnome.org/show_bug.cgi?id=628486